Move ValidationStructure to autogluon.common
CVSplitter -- which consumes the splits this class produces -- already lives in
autogluon.common, so producer and consumer sat on opposite sides of a package
boundary. Nothing in ValidationStructure is core-specific: it is a dataclass over
pandas/numpy/sklearn returning index arrays.

The one blocker was its problem_type constants, since autogluon.common deliberately
does not depend on autogluon.core. Problem types are shared vocabulary rather than
core-internal names, so they move to a new autogluon.common.constants and
autogluon.core.constants re-exports them; every module importing them from core keeps
working unchanged. The core-internal names in that file (AG_ARGS, REFIT_FULL_NAME,
...) stay where they are.

Also documents group_time_on in TabularPredictor.fit's validation_structure entry,
which listed only the other three keys.

Verified: full import chain loads, both constant paths resolve identically, 200 tests
pass across the model/callback/registry suites, and split parity stays 39/39 from the
new location.
